Issue 28339 improve deterministic id api 2 (#29103)
### Proposed Changes
* We're making the DeterministicIdentifierAPI more resilient. in case of
an error generating the seed, we return an empty Optional
* If an empty optional gets returned that means we failed to create the
seed therefore we fall back to the classic UUID generator
* The improvements in the DeterministicIdentifierAPI now guarantee that
a contentlet created with the same properties will have the same id even
after it gets deleted. There was a test saving identifier for later
clean-up that required some updates as the same identifier was being
sent for delete more than once.