build: Fixup developer Native CLI Build and upgrade Graal (#29110) (#29111)
Now that Quarkus for the CLI is on 3.5 it supports the Java 21 version
of GraalVM. Note the version numbering has changed and now the Major
version reflects the Java version. These versions are supported in
sdkman which makes the installation process easier. We do not want to
enforce use of the GraalVM for the full compilation and run of the
project but we can either compile just the cli module when generating
the native binary with GraalVM or setting the GRAALVM_HOME environment
property and Quarkus will use it when needed. sdkman automatically
installs the required native executable and makes it available on the
path.

The current version of GraalVM for java 21 can be installed with

sdk install java 21.0.2-graalce

changing to use it after install can be achieved with

sdk use java 21.0.2-graalce

or we can set the GRAALVM_HOME from the returned path by sdkman (e.g.
/Users/stevebolton/.sdkman/candidates/java/21.0.2-graalce) with

export GRAALVM_HOME=$(sdk home java 21.0.2-graalce)

It does not matter if we are still Java 11 the main Java version at the
moment

Tasks

1. Fix compilation issue on m1 mac by adding
-Wno-nullability-completeness compiler flag
2. Update justfile to allow for simple native build and test options
3. Change quarkus packaging type to uber-jar on the main cli module by
default and remove the need for a separate profile.

CI Build will be updated separately to the updated GRAALVM

Tests

- Test the just commands on different environments
- Test the native build verify to ensure that tests are run against the
native executable and an issue that only breaks the native build is
detected. Can break the native build by removing the build.properties
from tools/dotcms-cli/cli/src/main/resources/resources-config.json.



### Additional Info
Related to #29110 (Fixup developer Native CLI Build and upgrade Graal).


### Screenshots
Original             |  Updated
:-------------------------:|:-------------------------:
** original screenshot **  |  ** updated screenshot **