fix(uve): Resolve Authorization Issue in dotCMS Angular Headless Implementation #28949  (#28982)
### Proposed Changes
* Changed the strategy of hydration layout component in UVE: Now we
reuse the current dotcms session token to get the pageAPI responses, and
the layout component is hydrated by UVE using postMessage.
* Removed onReload functions: Now we dont need to "recharge" the iframe
after add/edit/etc contentlets inside UVE
* Added vanityUrl inside UVE
* Fixed vanityUrl in NextJS example 
* Fixed vanityUrl in Angular example
* This change introduces a new flow on the pageAsset fetch:

- Outside editor: We use the client sdk to get and hydrate the layout
component/header/footer
- Inside editor: We use also the page data from the UVE, and the data is
sent to the iframe and the component is re-hydrated.


Considerations:
-With this change, the website have a frontend user permission in SPA,
so pages like "members" cannot render in frontend page, because it have
a specific user permission.
To render it, we need to manually add the frontend permission


https://github.com/dotCMS/core/assets/144152756/6e401349-d3ef-4af7-ac64-fd249024602d

---------

Co-authored-by: KevinDavilaDotCMS <codeabien@gmail.com>
Co-authored-by: KevinDavilaDotCMS <kfariid@gmail.com>