ci: Fix default runid for base and ci artifacts when running on master branch. (#29223)
### Proposed Changes
* When running on master branch we are either using maven artifacts
built in the merge queue or from the current run id. CLI artifacts are
currently always generated in the current build and are not pulled from
the merge queue.
* There was an issue that the correct run-id was not being passed for
the sonarqube step or to pull maven repo for cli native builds.
* This fix makes sure the correct run-id is consistently passed through
and used by CLI and jobs requiring core artifacts.


### Checklist
- [ ] Tests
- [ ] Translations
- [ ] Security Implications Contemplated (add notes if applicable)

### Additional Info
** any additional useful context or info **

### Screenshots
Original             |  Updated
:-------------------------:|:-------------------------:
** original screenshot **  |  ** updated screenshot **