Avoid Layout bug when move Containers inside a Layout (#29227)
If we have a Template shared between several pages and then change the
Layout for one of these pages, a new Layout is created while the
Template remains unchanged.

In the code at the following link, we are checking which pages are using
the Template that we are going to update and then updating the
MultiTree:


https://github.com/dotCMS/core/blob/master/dotCMS/src/main/java/com/dotmarketing/portlets/templates/business/TemplateAPIImpl.java#L1301

The issue arises when different pages are using the same Template, and
we change the Layout for one of these pages. This action creates a new
Layout. At this point, when we query for pages that are using the
Template, we get an empty collection because no pages are using the new
Template since it has just been created.

Remember a Layout is really a template the only different is that the
Layout is used just for one Page.

Proposed Changes
We changed the method to receive the oldTemplate and the pages Id what
need to we changed, on the case that it is a Layout we are going to pass
just one page's id


https://github.com/dotCMS/core/pull/29227/files#diff-c74cebd0e6210e5d55c1e765543514a52f84f3f312277ddd3c80d21680fb545cR388

passing the oldTemplate allow us to get the oldLayout to compare the
changes so if the page was using a Template and now it is going to be
change by a Layout then we need the Template to compare.