chore(edit-content): fix jump with AI dialog (#29343)
### Parent Issue

https://github.com/dotCMS/core/issues/29173

### Proposed Changes
* Add hack CSS to avoid this jump

### Checklist
- [x] Tests
- [x] Translations
- [x] Security Implications Contemplated (add notes if applicable)

### Problem Statement

There is a small jump when you open the ai image dialog from the binary
field, do not happen from the block editor.


https://github.com/dotCMS/core/assets/31667212/4d27361e-4ed3-4011-a49f-6b55be97e1bc

### Additional info

This is a problem with PrimeNG v15 when we using Dialog + Accordions,
this issue is reported
[here](https://github.com/primefaces/primeng/issues/9453).

The problem happens when an accordion component is inside a p-dialog all
options are opened and then a few seconds later the options are closed
generating a small jump in the UI, here is a video with the problem.


https://github.com/user-attachments/assets/b92a9675-ac15-4a72-b8de-f44d52fe59b3

According to the issue, a quick fix is to add the following css code to
avoid this jump.

```css
:host ::ng-deep {
    .p-accordion-tab:not(.p-accordion-tab-active) {
        .p-toggleable-content {
            height: 0;
            overflow: hidden;
        }
    }

    .p-accordion-tab.p-accordion-tab-active {
        .p-toggleable-content {
            height: auto;
            overflow: auto;
        }
    }
}
```

As a result.


https://github.com/user-attachments/assets/6b393701-15e0-4774-958e-d3d70aafc084