#28813: Moving OpenAI Model paramenters to dotAI App configuration (#29236)
Removing hardcoded OpenAI models at enum class OpenAIModel. Instead they
are now part of the `dotAI.yml` application descriptor so the user can
be the one who configures them not only one but multiple models for
`text`, `image` and `embeddings`.
The way to specify more than one is to provide a comma delimited list in
the new dotAI App params.
Sometimes we accept a model to use in the payload of our AI endpoints,
for this matter we will validate that model and if it's invalid we will
throw an exception.
When is not present in the payload, then our backend will inject the
current model.
Which leads us to the question: How will the current model be
determined?
This is part of the work defined for
https://github.com/dotCMS/core/issues/29284 (model fallback)