fix: developer builds fail on yarn install when run mul (#29406) (#29411)
This change updates the commit hook and ensures that node_modules is not
processed by nx or eslint.

Related to #29406 (developer builds fail on yarn install when run mul).

We could have the maven command also run the yarn install and even the
lint and format to keep it totally consistent, but pulling out these
subsequent steps makes these node commands more visible in the
developer's console

### Testing

To test the commit hook without creating a commit you can modify the
exit code to ensure it always returns a non-zero code, also you can
create a test commit and then revert it with "git reset HEAD~1". or "git
reset --hard HEAD~1" if you do not want to keep any of the changes in
the test commit. Make sure you do not push with these changes

This change may run slower the first time if maven or yarn install has
not already been run to download the dependencies the first time.