fix(GraphQL) Refs #29289 (#29351)
### Proposed Changes
* This solves a problem that breaks the GraphQL API when removing
non-required fields from the CT the Schema becomes invalid and throws an
error.
* The error looks more or less like this: 
```
   object type 'Page2024' does not implement interface 'PageBaseType' because field 'seokeywords' is missing
   object type 'Page2024' does not implement interface 'PageBaseType' because field 'pagemetadata' is missing
```

Also, fix an issue that prevents the user from seeing the real error
thrown by the GraphQL Servlet
There is an issue with our GraphQL impl where the library code uses a
`NOPLogger` probably due to a conflict in the slf4j dependency. with
this fix, we use our code to dispatch and process the request and handle
and report the errors

The postman was adjusted not to fail. Because this new restriction
creating the schema will leave out some the non-required fields from the
BaseType collections