fixes #27978 chore(Language Variables) : Register the new Locale Variable Upgrade Task as a Data Task (#29433)
### Proposed Changes
* Registers the `Task240306MigrateLegacyLanguageVariables` Data Task
which migrates the legacy Language Variables living in the .properties
files under the `/assets/messages/` folder into Contentlets of type
`Language Variable`.
* As specified in [this comment in the
ticket](https://github.com/dotCMS/core/issues/27978#issuecomment-2248749304),
the Language Variables will be created for the specific base language if
it exists. If it doesn't, the entry will be created for every Locale
with that specific base language.
* The Language Variable Contentlets will be published after the data
task has run.
* If a key coming from the .properties file already exists as a
Contentlet, it will be ignored and the existing Contentlet will be kept
as is.