#29281: adding a centralized OpenAI api-key validation procedure (#29420)
Avoiding sending requests when there is no OpenAI API key to use.
Putting it in a centralized place.
Adjusting a couple of tests to reflect this logic change.
Also since from the provided logs there was an Jackson parsing error,
new fields were added to "DTO" class to consider the error field.
Initially when implementing IT for AI classes, I took the approach of
mocking or building an `AppConfig` class on the flight instead of
creating an `dotAI` app secrets so config classes can be created out of
them. Just as it happens at production code. I'm introducing some
changes to correct this as well.