Issue 29293 cannot change folder url to uppercase (#29450)
When renaming a folder an insert statement was being executed to update
the folder. But when the name is the same (lowercase or uppercase) that
insert was throwing a _duplicate key value violates unique constraint
"idx_ident_uniq_asset_name"_

Now the validation to check if it is renamed or not is
_equalsIgnoreCase_ so in these cases the name will be considered as the
same, and update the folder in a correct way.

Co-authored-by: erickgonzalez <erick.gonzalez@dotcms.com>