Issue 28779 exception ss portlet (#29458)
I understand the frustration of dealing with the SS portlet becoming
unusable when the default index is removed. I'm glad to share that with
this change, even if the index does not exist, a warning will be logged
and the portlet will render successfully. This means you'll be able to
set another SS index as the default one without any issues.

A test was added.