fix(design system): Autocomplete and Tags should show overflow for long tags (#29511)
### Proposed Changes
* Fix CSS to add ellipsis and overflow 

### Note 

The fix is not as good as it should be, because this tags component is
created and managed by a really old JS code that is not building the tag
as it should in a semantic way, this means that is really difficult to
add styles to this component and I tried to make it more easier by
adding a `span` to the component but it broke all the functionality of
adding/removing tags.

I don't think is worth a refactor since is legacy code and we are
already in the way to use the new Edit Content.

So, I had to make CSS tricks to make it work as expected.

### Screenshots



https://github.com/user-attachments/assets/c23b1ec0-e90b-4aae-a5db-d7b9b445f869