fix(api): Regression new custom portlets created without vie (#29524) (#29525)
### Proposed Changes
* Make sure the initParams for the DotPortlet builder include the params
from the base "content" portlet
* Fixup PortletResource.json "Retrieve an existing Portlet" postman test
to ensure view-action is set to /ext/contentlet/view_contentlets

To Replicate

Replicate issue by trying to add a custom portlet from the UI with the
"Add to Menu" option from the content type. The Menu is created but
produces a blank page and an error from the PortletRequestProcessor
"{{portlet name}} does not have any paths specified" is returned
Looking at the XML created in the portlets table it is missing the
<init-param><name>view-action</name><value>/ext/contentlet/view_contentlets</value></init-param>
The response from 

Also when running the postman test Portlet Resource, Create a Custom
Portlet followed by Retrieve an existing Portlet shows the init-params
missing the required view-action init-param. The test passes though as
this parameter is not verified

Test 
* When adding a custom portlet from the UI "Add to Menu" ensure that the
xml correctly containst the
<init-param><name>view-action</name><value>/ext/contentlet/view_contentlets</value></init-param>
* Ensure that the custom potlet page loads and not a blank page and
there is not error from the PortletRequestProcessor "{{portlet name}}
does not have any paths specified"


### Checklist
- [ ] Tests
- [ ] Translations
- [ ] Security Implications Contemplated (add notes if applicable)

### Additional Info
Related to #29524 (Regression new custom portlets created without vie).


### Screenshots
Original             |  Updated
:-------------------------:|:-------------------------:
** original screenshot **  |  ** updated screenshot **