Fixing category children counting (#29566)
Issue: Counting Category Children in the Tree Table

We encountered a problem while counting the child categories in our tree
table. This table uses child and parent fields to define the
parent-child relationships between categories.

The query used for counting was:

`SELECT COUNT(*) FROM tree WHERE parent = ?`

However, this table is also used to store relationships between
categories and content. In these cases, the category is saved as the
parent and the content as the child. As a result, the query above
mistakenly counts content items as child categories, leading to
incorrect counts.

### Proposed Changes
* Change the SQL query used to count the children to discard the case
when the child is really a Contentlet


https://github.com/dotCMS/core/pull/29566/files#diff-0e6819b564f7ed83bb6feb79c419f56b0470e9d6dcf685c0af460103ee87408cR30