chore: Refactor github workflows to group key top level w (#28554) (#29506)
* Change file naming convention for actions and workflows.
* Workflows do not allow subfolders so file naming prefix used to
provide structure.
* Clearly separate new core cicd workflows and actions from workflows
related to Issue management and non cicd processes, categorize existing
legacy release workflows that have not been migrated yet to prevent
confusion and be clear on work to be done.
* Add README files and documentation on the whole workflow

Impacts
Although this is mostly a non functional change with just the file paths
and references to them updated as well as the workflow titles to be
consistent there are a couple of things to consider.
1. This updates the badges on the project root README to the new
workflow filenames. These will remain broken links though until the code
goes through the merge queue and into the nightly build.
2. Due to the change in workflow name the old and new workflow names
will be grouped separately in the Actions page, Once the PR is merged we
should replace the bookmarked workflow.

This PR resolves #28554 (Refactor github workflows to group key top
level w).