Fix/Feat (CLI/Core) Cleanup content type descriptor (#29579)
This PR includes changes to make the CLI Content Type descriptor more
friendly for the end user:
1. CLI: Now we display, for Relationship fields, an enum name for the
relationship cardinality instead of a number. For example: MANY_TO_MANY,
ONE_TO_ONE, etc.
2. CLI: The `systemActionMappings` is simplified by removing unnecessary
properties and using the format that it is expected by the Content Type
resource, example:
```json
  "systemActionMappings" : {
    "NEW" : "b9d89c80-3d88-4311-8365-187323c96436",
    "EDIT": "ceca71a0-deee-4999-bd47-b01baa1bcfc8",
    "UNPUBLISH": "38efc763-d78f-4e4b-b092-59cd8c579b93",
    "UNARCHIVE": "c92f9aa1-9503-4567-ac30-d3242b54d02d"
  }
```

Also, during the "cleanup" process the logic in the Content Type
Resource related to the system action mappings was improved allowing to
properly to add, edit and remove system action mappings.

---------

Co-authored-by: Nollymar Longa <nollymar.longa@dotcms.com>