Issue 29259  (#29656)
The issue was being generated because of the addition of an image as
dotAsset, if it was a File Asset, then everything was working ok.

- To solve the removal of the image a check was added to see if it is a
dotAsset, then avoid to cache the data.
- To solve the missing data in the related content, "fileAsset" was
added into the dotAsset properties with the data that "asset" property
had, this workaround is to make the mapping of the fields like a File
Asset.