Fix (CLI/Core) Fixed flakey test (#29723)
In this PR, we removed an unnecessary lookup that, under very rare
circumstances, was returning an outdated cached version of the Content
Type. I believe this happened because the cache invalidation process was
still in progress. My theory is that the invalidation process isn’t
always immediate, and the lookup occurred before it was completed,
leading to the retrieval of an outdated version of the Content Type.

I managed to isolate the issue, and in the few instances where it was
reproduced, allowing enough time before calling `contentTypeAPI.find`
resulted in the correct version being returned. This observation led me
to conclude that the issue was indeed related to cache invalidation
timing.