fix: revert session manager jar location change from PR #29694 (#23631) (#29746)
### Proposed Changes
* Fix classloader issue with session manager introduced in PR #29694
* Use CATALINA_HOME in setenv scripts instead of TOMCAT_HOME. The first
is always available, the second is only set within the docker startup
scripts

Instructions for Log4j require the classes to be added to the system
classloader when tomcat loads by adding to the main java CLASSPATH.
These logging jars do not need to reference any of the tomcat jars
the session manager jars do need to reference the tomcat jars e.g.
catalina.jar in the main lib folder. This is what is called the
common.loader and is defined in catalina.properties file.
the session manager jars either need to be in a folder referenced in
this common.loader property or as before dropped into the existing lib
folder. There is no need to change the old behaviour at this time so we
can just change the install location back.