ci: Fix SonarQube timeout property (#29886) (#29887)
### Proposed Changes
* Fix the action to the current commit
https://github.com/sonarsource/sonarqube-quality-gate-action/tree/dc2f7b0dd95544cd550de3028f89193576e958b9
for reproducability instead of auto updting any change to master branch.
Includes the timeout change in this PR
https://github.com/SonarSource/sonarqube-quality-gate-action/pull/50
*
* set the new pollingTimeoutSec to 600 (e.g. 10 minutes)

To test, check the logs of the build run in the sonarqube step and it
should show the timeout passed in as "600" instead of "300"

### Checklist
- [ ] Tests
- [ ] Translations
- [ ] Security Implications Contemplated (add notes if applicable)

### Additional Info
This PR resolves #29886 (Fix SonarQube timeout property).


### Screenshots
Original             |  Updated
:-------------------------:|:-------------------------:
** original screenshot **  |  ** updated screenshot **