fix(uve): Enterprise features need to be disabled on community license (#30025)
### Proposed Changes
* Updated the logic to disable navigation items based on whether they
require an enterprise license and the current license status.

### Checklist
- [x] Tests
- [x] Translations
- [x] Security Implications Contemplated (add notes if applicable)

### Additional Info
In this update, navigation items that are enterprise (layout, rules,
experiments) will be disabled if the current license is not an
enterprise license. This ensures that enterprise-only features are not
accessible in the Community Edition, improving the user experience by
preventing confusion and frustration.

### Screenshots
#### Editing a page in community edition

![image](https://github.com/user-attachments/assets/4e30c553-db35-40cd-bc34-686d34286d21)

#### Editing a page in enterprise edition

![image](https://github.com/user-attachments/assets/dfead069-95aa-4a6d-81b3-c736cef40556)

This pr fixes #29977