fix(deployment): Update pom file with the latest empty starter (#29642) (#29718)
### Proposed Changes
**update-pom**:

It updates the `pom.xml` file with the latest version of the `Empty
Starter`.
A new branch is created, the version in pom.xml is updated, and a Pull
Request (PR) is automatically generated to merge these changes into the
main branch.
The PR URL is captured as an output for further notification steps.

**send-notification**:

If an Empty Starter was updated, the notification message also includes
a link to the PR that requires approval.



Related to #29642 (Update pom file with the latest empty starter).