Issue 29238 Attaching files feature on tasks is not properly working (#29992)
When attaching dot assets it was throwing an exception that the file
wasn't a file asset, so I implemented a conversion from dotAsset to
fileAsset, first made the _convertToFileAsset_ function public and
static to be accessible from outside its class and to reuse the logic of
that function.
Then implemented the extra logic to convert the dotAsset to fileAsset in
a new function _transformDotAsset_, where also a new property "fileLink"
is added to then display it in the href of the jsp, this way the file
can be accessible without problem when clicking on it.

---------

Co-authored-by: erickgonzalez <erick.gonzalez@dotcms.com>