refactor(deployment): Update file references from main to master (#29841) (#30058)
### Proposed Changes
* All the references to master branch of the project has been changed to
main.

The switch from "master" to "main" as the default branch name in Git is
part of an effort to use more inclusive and neutral language. The term
"master" can carry negative historical connotations, while "main" is a
clearer and more descriptive term that reflects its role as the primary
branch. Updating these references aligns the project with modern best
practices in the tech community for creating an inclusive environment.

### Additional Info
Related to #29841 (Update file references from main to master).


### Screenshots
Original             |  Updated
:-------------------------:|:-------------------------:
** original screenshot **  |  ** updated screenshot **

---------

Co-authored-by: Nollymar Longa <nollymar.longa@dotcms.com>