fix(backend):  Unnecessary additional Host caching in resolveHostName causing null return (#30108) (#30132)
There is confusion in the code whether we should be caching the mapping
between requested host and the backend host, or whether the mapping
should cache the default entry returned.

The issue is occuring as a request for "localhost" that does not appear
in the db hosts used to be mapped to the default host in
com.dotmarketing.portlets.contentlet.business.HostAPIImpl#resolveHostName
so null was not returned from this method.

With the change to this method in the original PR when it calls 


com.dotmarketing.portlets.contentlet.business.HostAPIImpl#resolveHostNameWithoutDefault
it ends up with a 404 entry in the host cache for localhost. A short
time later it should replace this at the end of the method by adding a
mapping to the resolved default host, subsequent calls to localhost
should then end up using this cached entry to the specific default host
that was cached.

There is a short window of time though where the 404 entry was added,
but the cache has not been updated to the default host where the current
logic on finding the 404 entry is to just return null.
This null was causing the ThreadNameFilter to throw a null pointer
exception as it is using the result without null checking.

Caching the resolved default host may also cause issues if the default
host is changed as it may not take this change into account.

There should actually be no reason to handle the caching logic in
resolveHostName and only have it within resolveHostNameWithoutDefault
that it calls. This way even if there is a 404 entry for an unknown like
"localhost" it will always return the current default value (that should
itself be cached) and not a null.

We can therefore revert the original changes to just this one method but
still make use of the caching logic provided when it calls
resolveHostNameWithoutDefault ensuring null is not returned by
resolveHostName method.


This PR resolves #30108 (Flakey CLI Test
PushServiceIT.Test_Push_New_Site).