fix(UVE): make changes to layout you lose container editable areas (#30189)
### Proposed Changes
* Replaced `updatePageResponse` with the `reload` method in
`EditEmaLayoutComponent` to trigger a page reload after saving.

### Checklist
- [x] Tests
- [x] Translations
- [x] Security Implications Contemplated (add notes if applicable)

### Additional Info
Issue #30116 was identified due to an issue with the response from the
backend when making a POST request to `api/v1/page/{pageId}/layout`. The
response contains the `page.rendered` object, which is used to patch the
UVE page state. However, this `page.rendered` HTML comes in **preview
mode**, causing empty layout containers to disappear, making it
impossible to edit the page correctly.

### Screenshots

https://github.com/user-attachments/assets/86e03108-308d-40c4-bea2-bafe8b787a15

This pr fixes #30116