chore(dotcms-ui): Fix error with autocomplete #30249 (#30269)
### Parent Issue
#30249 

### Proposed Changes
* Change singal approach

### The problem

This computed signal caused the error:

```ts
$filteredSuggestions = computed(() => {
    const classes = this.$classes();
    const query = this.$query();

    if (!query) {
        return classes;
    }

    return classes.filter((item) => item.includes(query));
});

filterClasses({ query }: AutoCompleteCompleteEvent): void {
    this.$query.set(query);
}
```

The signals functionality in PrimeNg doesn't works very well. In the
previous signal, `$filteredSuggestions` doesn't trigger a change if the
classes or query are the same as the current state. This is desirable.
However, PrimeNg
[waits](https://github.com/primefaces/primeng/blob/d6c7ffcfc5cb9a8ed355399ddae276c279836ccf/src/app/components/autocomplete/autocomplete.ts#L541)
for a change in the `suggestions` input after the `completeMethod` is
called to turn off the loading. This is why the loading indicator
remains active but never deactivates.

To address this, we manually forced a change in `$filteredSuggestions`
and used `[...filteredClasses]` to create a new array.

```ts
filterClasses({ query }: AutoCompleteCompleteEvent): void {
      const classes = this.$classes();
      const filteredClasses = query ? classes.filter((item) => item.includes(query)) : classes;
      this.$filteredSuggestions.set([...filteredClasses]);
  }
```

### Checklist
- [x] Tests
- [x] Translations
- [x] Security Implications Contemplated (add notes if applicable)

### Screenshots



https://github.com/user-attachments/assets/5a199569-9192-4acb-89bd-3fb6530aefce